qcacld-3.0: Align score boost percent by MLO type with F/W
Align MLO score boost percent by MLO type with F/W.
MLSR boost 1%
EMLSR boost 2%
MLMR boost 3%

CRs-Fixed: 3912085

+++ components/umac/mlme/connection_mgr/dispatcher/inc/wlan_cm_roam_public_struct.h
@@ -1007,7 +1007,7 @@ struct scoring_param {
 	struct per_slot_score oce_wan_scoring;
 #ifdef WLAN_FEATURE_11BE_MLO
 	uint8_t eht_caps_weightage;
-	uint8_t mlo_weightage;
+	uint32_t mlo_weightage;
 #endif
 	int32_t security_weightage;
 	uint32_t security_index_score;
